web/realtime-engine-websocket/requirements.txt
aiohttp-session==1.0.1
aiohttp==2.3.1
aiopg==0.13.1
aiohttp-devtools
python-engineio
python_socketio

